Extract platform check utility to common utilities
